refactor: Move more "ctx" traits into authly-domain
