+++ lib/authly-domain/Cargo.toml
@@ -15,11 +15,27 @@ doctest = false
 [dependencies]
 authly-db = { path = "../authly-db" }
 authly-common.workspace = true
+aes-gcm-siv = "0.11"
+anyhow = "1"
+arc-swap = "1.7"
+blake3 = "1.5"
 fnv = "1"
 hexhex = "1"
+indexmap = "2.7"
 int-enum = "1"
+jsonwebtoken = "9"
 itertools = "0.14"
+pem = "3"
+rcgen = { git = "https://github.com/rustls/rcgen.git" }
+reqwest = { version = "0.12", default-features = false, features = [
+    "http2",
+    "json",
+    "rustls-tls",
+] }
+rustls = { version = "0.23", default-features = false }
 thiserror = "2"
 time = "0.3"
 tokio = { version = "1", features = ["macros"] }
 tracing = "0.1"
+x509-parser = "0.16"
+zeroize = "1.8"
