+++ lib/authly-domain/src/lib.rs
@@ -1,4 +1,8 @@
 pub mod builtins;
+pub mod cert;
 pub mod ctx;
 pub mod directory;
+pub mod encryption;
 pub mod id;
+pub mod instance;
+pub mod tls;
