+++ src/authority_mandate/submission/authority.rs
@@ -1,17 +1,18 @@
 //! Submission, authority side
 
 use authly_common::id::ServiceId;
-use authly_domain::ctx::GetDb;
+use authly_domain::{
+    cert::authly_ca,
+    ctx::{GetDb, GetInstance},
+    tls::{AuthlyCert, AuthlyCertKind},
+};
 use rand::{rngs::OsRng, Rng};
 use rcgen::{CertificateSigningRequestParams, DnValue, PublicKeyData};
 use tracing::warn;
 
 use crate::{
     audit::Actor,
-    cert::authly_ca,
-    ctx::GetInstance,
     db::authority_mandate_db::{self, AmDbError},
-    tls::{AuthlyCert, AuthlyCertKind},
     util::serde::UrlSafeBase64,
 };
