+++ src/bus/handler.rs
@@ -1,9 +1,9 @@
 use anyhow::Context;
-use authly_domain::ctx::GetDb;
+use authly_domain::ctx::{GetDb, GetDecryptedDeks, RedistributeCertificates, SetInstance};
 use tracing::info;
 
 use crate::{
-    ctx::{ClusterBus, GetDecryptedDeks, RedistributeCertificates, ServiceBus, SetInstance},
+    ctx::{ClusterBus, ServiceBus},
     db::{
         cryptography_db::load_authly_instance,
         directory_db::{self, query_dir_key},
