+++ src/k8s/k8s_auth_server.rs
@@ -5,7 +5,11 @@ use std::{
 
 use anyhow::anyhow;
 use authly_common::id::ServiceId;
-use authly_domain::ctx::{GetBuiltins, GetDb};
+use authly_domain::{
+    cert::{client_cert, server_cert, Cert, CertificateParamsExt},
+    ctx::{GetBuiltins, GetDb, GetInstance},
+    instance::AuthlyInstance,
+};
 use axum::{body::Bytes, extract::State, response::IntoResponse, routing::post, Extension};
 use axum_extra::{
     headers::{authorization::Bearer, Authorization},
@@ -21,10 +25,7 @@ use rustls::{pki_types::PrivateKeyDer, ServerConfig};
 use tracing::{error, info};
 
 use crate::{
-    cert::{client_cert, server_cert, Cert, CertificateParamsExt},
-    ctx::GetInstance,
     db::service_db,
-    instance::AuthlyInstance,
     util::remote_addr::{remote_addr_middleware, RemoteAddr},
     AuthlyCtx, EnvConfig,
 };
