+++ src/service.rs
@@ -2,10 +2,10 @@
 
 use authly_common::id::ServiceId;
 use authly_db::DbError;
-use authly_domain::ctx::{GetBuiltins, GetDb};
+use authly_domain::ctx::{GetBuiltins, GetDb, HostsConfig};
 use tracing::info;
 
-use crate::{ctx::HostsConfig, db::service_db};
+use crate::db::service_db;
 
 pub async fn get_service_hosts(
     deps: &(impl GetDb + GetBuiltins + HostsConfig),
