+++ src/test_support/test_ctx.rs
@@ -9,7 +9,15 @@ use authly_common::id::ServiceId;
 use authly_db::sqlite_pool::{SqlitePool, Storage};
 use authly_domain::{
     builtins::Builtins,
-    ctx::{GetBuiltins, GetDb},
+    cert::{authly_ca, client_cert, key_pair},
+    ctx::{
+        Directories, GetBuiltins, GetDb, GetDecryptedDeks, GetHttpClient, GetInstance, HostsConfig,
+        KubernetesConfig, LoadInstance, RedistributeCertificates, SetInstance,
+    },
+    directory::PersonaDirectory,
+    encryption::DecryptedDeks,
+    instance::{AuthlyId, AuthlyInstance},
+    tls::{AuthlyCert, AuthlyCertKind},
 };
 use indexmap::IndexMap;
 use tokio_util::sync::{CancellationToken, DropGuard};
@@ -22,19 +30,12 @@ use crate::{
         service_events::{ServiceEventDispatcher, ServiceMessageConnection},
         BusError,
     },
-    cert::{authly_ca, client_cert, key_pair},
-    ctx::{
-        ClusterBus, Directories, GetDecryptedDeks, GetHttpClient, GetInstance, HostsConfig,
-        KubernetesConfig, LoadInstance, RedistributeCertificates, ServiceBus, SetInstance,
-    },
+    ctx::{ClusterBus, ServiceBus},
     db::{
         cryptography_db,
         init_db::{self},
     },
-    directory::PersonaDirectory,
-    encryption::{gen_prop_deks, DecryptedDeks, DecryptedMaster},
-    instance::{AuthlyId, AuthlyInstance},
-    tls::{AuthlyCert, AuthlyCertKind},
+    encryption::{gen_prop_deks, DecryptedMaster},
     IsLeaderDb, Migrations,
 };
