+++ src/tests/test_authority_mandate.rs
@@ -1,5 +1,9 @@
 use authly_common::id::PersonaId;
 use authly_connect::TunnelSecurity;
+use authly_domain::{
+    cert::{server_cert, CertificateParamsExt},
+    ctx::GetInstance,
+};
 use itertools::Itertools;
 use rcgen::CertificateSigningRequestParams;
 use test_log::test;
@@ -16,8 +20,6 @@ use crate::{
             mandate_identity_signing_request,
         },
     },
-    cert::{server_cert, CertificateParamsExt},
-    ctx::GetInstance,
     proto::mandate_submission::AuthlyMandateSubmissionServerImpl,
     tests::{rustls_server_config_no_client_auth, spawn_test_connect_server, TestCtx},
 };
