+++ src/tls.rs
@@ -1,64 +1,15 @@
-use std::fmt::Display;
-use std::str::FromStr;
 use std::sync::Arc;
 
 use anyhow::anyhow;
-use authly_common::id::ServiceId;
+use authly_domain::{
+    cert::{server_cert, CertificateParamsExt},
+    ctx::GetInstance,
+};
 use futures_util::StreamExt;
-use pem::{EncodeConfig, Pem};
-use rcgen::CertificateParams;
-use rustls::pki_types::CertificateDer;
 use rustls::{pki_types::PrivateKeyDer, server::WebPkiClientVerifier, RootCertStore, ServerConfig};
 use tracing::info;
 
-use crate::cert::{server_cert, CertificateParamsExt};
-use crate::ctx::GetInstance;
 use crate::{AuthlyCtx, AuthlyInstance};
-
-#[derive(Clone, Debug)]
-pub struct AuthlyCert {
-    pub kind: AuthlyCertKind,
-    pub certifies: ServiceId,
-    pub signed_by: ServiceId,
-    pub params: CertificateParams,
-    pub der: CertificateDer<'static>,
-}
-
-impl AuthlyCert {
-    pub fn certificate_pem(&self) -> String {
-        pem::encode_config(
-            &Pem::new("CERTIFICATE", self.der.to_vec()),
-            EncodeConfig::new().set_line_ending(pem::LineEnding::LF),
-        )
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub enum AuthlyCertKind {
-    Ca,
-    Identity,
-}
-
-impl Display for AuthlyCertKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ca => write!(f, "CA"),
-            Self::Identity => write!(f, "identity"),
-        }
-    }
-}
-
-impl FromStr for AuthlyCertKind {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "CA" => Ok(Self::Ca),
-            "identity" => Ok(Self::Identity),
-            _ => Err(()),
-        }
-    }
-}
 
 pub fn init_tls_ring() {
     let _ = rustls::crypto::ring::default_provider().install_default();
