+++ src/web/auth/oauth.rs
@@ -1,7 +1,10 @@
 use std::{borrow::Cow, collections::BTreeMap};
 
 use anyhow::{anyhow, Context};
-use authly_domain::ctx::GetDb;
+use authly_domain::{
+    ctx::{Directories, GetDb, GetDecryptedDeks, GetHttpClient},
+    directory::{OAuthDirectory, PersonaDirectory},
+};
 use axum::{
     extract::{FromRef, Path, Query, State},
     response::{IntoResponse, Response},
@@ -13,8 +16,6 @@ use reqwest::Url;
 use tracing::warn;
 
 use crate::{
-    ctx::{Directories, GetDecryptedDeks, GetHttpClient},
-    directory::{OAuthDirectory, PersonaDirectory},
     persona_directory::{self, ForeignPersona},
     session::init_session,
     util::base_uri::ProxiedBaseUri,
